requirements.txt
PyYAML
keyboard>=0.13.5
pynput>=1.7.6
psutil>=5.9.0
pandas>=1.3.0
scikit-learn>=1.0.0
xgboost>=1.5.0
joblib>=1.0.0
numpy>=1.21.0
pywebview>=5.0
python-dotenv>=1.0.0
requests>=2.28.0
keyring>=24.0.0
google-generativeai>=0.8.0
